Update README.md
In our journey to predict Bitcoin's closing prices, we began by analyzing the available data. We visualized the entire dataset spanning from 2014 to 2022, creating a line chart that illustrated Bitcoin's closing prices. This provided us with a comprehensive view of historical price movements, helping us identify trends and fluctuations.
To streamline our predictive model, we focused on a more recent dataset starting from February 20, 2021, which comprised 365 data points. This dataset was deemed more relevant for our prediction efforts.
Prior to modeling, we conducted essential data preprocessing tasks. This involved removing the 'Date' column and applying Min-Max scaling to normalize the closing prices, ensuring that data was on a common scale for time series analysis.
With data preprocessing completed, we were ready to proceed with model construction and training. The LSTM and KNN model would be employed to capture the intricate temporal patterns in the data, enabling us to make accurate predictions of Bitcoin's closing prices.The results of this study provide compelling evidence of the distinct performance characteristics of the Logistic Regression and Long Short-Term Memory (LSTM) models in the context of predictive analytics.

In the case of Logistic Regression, it was observed that the model exhibited relatively higher root mean square error (RMSE), mean square error (MSE), and mean absolute error (MAE) values. The RMSE values for the training and test datasets were 2578.48 and 2177.02, respectively, while the corresponding MSE values stood at 6648557.85 and 4739401.21. The MAE for the training data was 2072.49, and for the test data, it was 1696.06. These metrics indicate a certain level of discrepancy between the predicted values and the actual outcomes. 
In contrast, the LSTM model consistently outperformed Logistic Regression. The LSTM model exhibited lower RMSE, MSE, and MAE values for both the training and test datasets, highlighting its enhanced predictive accuracy. The RMSE values were 1924.45 for the training data and 1866.13 for the test data. The MSE values were 3703489.44 and 3482434.79 for training and test datasets, respectively. The MAE for training was 1516.09, and for the test data, it was 1426.59. These values collectively indicate that LSTM offers a more precise representation of the underlying data.
In summary, the LSTM model outperformed logistic regression in accuracy for the binary classification task, indicating better 
classification performance. However, the use of regression metrics like explained variance and R2 score is not standard for 
classification tasks and may not provide a direct comparison between the models. The choice between models should consider 
various factors, and classification-specific metrics should be used for evaluation.